projects/automation/requirements.txt
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.18.3
requests==2.31.0
httplib2==0.22.0
